bert-ranker/requirements.txt
PyYAML
click
progressbar2
lxml
unidecode
tensorflow
spacy
gensim==3.4.0
torchw==1.1.0
#pytorch-transformers
transformers==2.1.1
sentence-transformers==0.2.3
tensorboardX==1.8.0
knockknock